refactor(workbench): flatten the js/lattice folder into the resources root
