+++ workbench/resources/js/app.tsx
@@ -11,8 +11,8 @@ import {
 import { configureI18nFromPageProps, LocaleReload } from "@lattice-php/lattice/i18n";
 import { createRoot } from "react-dom/client";
 import sprite from "virtual:svg-sprite";
-import { appColumns } from "./lattice/columns";
-import { WORKBENCH_I18N_NAMESPACE } from "./lattice/i18n";
+import { appColumns } from "./columns";
+import { WORKBENCH_I18N_NAMESPACE } from "./i18n";
 
 const appRegistry = extendRegistry(registry, appColumns);
